Check tester usaito,TR42:all passed, pft:650/694

+++ include/formalize.h
@@ -6,7 +6,7 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/22 12:04:49 by kamitsui          #+#    #+#             */
-/*   Updated: 2023/03/24 11:30:38 by kamitsui         ###   ########.fr       */
+/*   Updated: 2023/03/29 22:43:39 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,9 +26,9 @@ void	adjust_string(const char *str, t_sm *machine);
 void	adjust_sign(char *str, t_sm *machine);
 void	adjust_hash(char *str, t_sm *machine);
 void	pad_str(const char *str, t_sm *machine);
-void	pad_n_str(const char *str, t_sm *machine, size_t limit);
-void	pad_width(t_sm *machine, size_t offset);
-void	pad_prec(t_sm *machine, size_t len);
+void	pad_n_str(const char *str, t_sm *machine, int limit);
+void	pad_width(t_sm *machine, int offset);
+void	pad_prec(t_sm *machine, int len);
 void	sequence_signed(const char *str, t_sm *machine);
 void	sequence_unsigned(const char *str, t_sm *machine);
