+++ src/formalize_utils/pad_n_str.c
@@ -6,7 +6,7 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/23 11:29:10 by kamitsui          #+#    #+#             */
-/*   Updated: 2023/03/24 11:11:29 by kamitsui         ###   ########.fr       */
+/*   Updated: 2023/03/29 22:49:11 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 #include "process.h"
 #include "libft.h"
 
-void	pad_n_str(const char *str, t_sm *machine, size_t limit)
+void	pad_n_str(const char *str, t_sm *machine, int limit)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < limit && (machine->state != ERROR))
