+++ src/formalize_utils/pad_prec.c
@@ -6,7 +6,7 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/23 14:27:06 by kamitsui          #+#    #+#             */
-/*   Updated: 2023/03/26 10:48:33 by kamitsui         ###   ########.fr       */
+/*   Updated: 2023/03/29 22:49:28 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 #include "process.h"
 #include "libft.h"
 
-void	pad_prec(t_sm *machine, size_t len)
+void	pad_prec(t_sm *machine, int len)
 {
-	size_t	i;
+	int	i;
 
 	if (machine->prec == 0)
 		return ;
